back/pyproject.toml
[tool.poetry]
name = "beck"
version = "0.1.0"
description = ""
authors = ["Miha <xd383727@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
pdoc3 = "^0.10.0"
Pillow = "^9.1.1"
djoser = "^2.1.0"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
